Allow even newer versions of signxml
### Description

Signxml 3.2 uses now-removed pyopenssl modules and methods (signxml.verifier
uses OpenSSL.crypto in particular). Signxml 4.0 is a big change under the hood
but the public API remains the same, except for use of ca_path, which we don't
appear to use anywhere. By updating this, we'll then be able to update pyopenssl
and cryptography.

### Testing

Very challenging to really test...

### Task and Issue Tracking

- [Asana Task](https://app.asana.com/0/1208949208553087/1208949208553092/f)